refactor: lift agentic mode to context_manager package; address PR feedback
- Move agentic impl out of conversation_manager/ into a top-level
  context_manager/modes/agentic/ package (Python and TS).
- pin_context: reject negative/out-of-range indices instead of wrapping.
- summarize_context/truncate_context: clamp keep_recent to the minimum.
- Trim verbose comment in _resolve_context_manager.
