+++ strands-py/src/strands/context_manager/__init__.py
@@ -1,0 +1,1 @@
+"""Context management modes that operate on an agent's conversation."""
