+++ strands-py/tests/strands/agent/conversation_manager/test_token_usage_middleware.py
@@ -6,7 +6,7 @@ import pytest
 
 from strands import Agent
 from strands._middleware.stages import InvokeModelContext
-from strands.agent.conversation_manager.modes.agentic.agentic_context import create_token_usage_middleware
+from strands.context_manager.modes.agentic.agentic_context import create_token_usage_middleware
 from strands.types.content import Message
 from tests.fixtures.mocked_model_provider import MockedModelProvider
